+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Set correct values for `allow_all`, `deny_all`, and `enforce`.
+
 ## [0.0.2]
 
 ### Fixed